Remove old CLI + cleanup (#1519)
* Remove old cli

* Update theme

* Remove shim

* display_version: True
