+++ nebari/initialize.py
@@ -49,14 +49,8 @@ def base_configuration():
                 "hub_title": None,
                 "hub_subtitle": None,
                 "welcome": None,
-                "logo": "/hub/custom/images/jupyter_nebari_logo.svg",
-                "primary_color": "#4f4173",
-                "secondary_color": "#957da6",
-                "accent_color": "#32C574",
-                "text_color": "#111111",
-                "h1_color": "#652e8e",
-                "h2_color": "#652e8e",
-                "version": f"v{__version__}",
+                "logo": "https://raw.githubusercontent.com/nebari-dev/nebari-design/main/logo-mark/horizontal/Nebari-Logo-Horizontal-Lockup-White-text.svg",
+                "display_version": True,
             }
         },
         "helm_extensions": [],
