+++ nebari/schema.py
@@ -542,8 +542,6 @@ class InitInputs(Base):
     kubernetes_version: typing.Union[str, None] = None
     ssl_cert_email: typing.Union[str, None] = None
     disable_prompt: bool = False
-    # TODO remove when Typer CLI is out of BETA
-    nebari: bool = False
 
 
 class Main(Base):
