+++ setup.cfg
@@ -65,7 +65,6 @@ dev =
 
 [options.entry_points]
 console_scripts =
-    # nebari = nebari.__main__:main
     nebari = nebari.cli.main:app
 
 [tool:pytest]
